requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
asgiref==3.6.0
assertpy==1.1
async-generator==1.10
attrs==23.1.0
backports.zoneinfo==0.2.1
behave @ git+https://github.com/behave/behave@ffd677601d1a997d94cb80e5308ef0cc065a1fe3
Cerberus==1.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver==2.24.1
colorama==0.4.6
config==0.5.1
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
crispy==0.7.4
crispy-bootstrap4==2022.1
cucumber-tag-expressions==4.1.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
Django==3.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-tinymce==3.6.1
django-user-sessions==2.0.0
dnspython==2.3.0
exceptiongroup==1.1.1
execnet==1.9.0
fabio==2023.6.0
filelock==3.12.0
fonttools==4.40.0
greenlet==2.0.2
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonpath==0.82
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.2
numpy==1.24.3
outcome==1.2.0
packaging==23.1
parse==1.19.0
parse-type==0.6.0
people==0.1.9
pep8==1.7.1
Pillow==9.5.0
pipenv==2023.4.29
pipfile==0.0.2
platformdirs==3.5.0
pluggy==1.0.0
pycparser==2.21
pyflakes==3.0.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PySocks==1.7.1
pytest==7.3.1
pytest-bdd==6.1.1
pytest-cache==1.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-flakes==4.0.5
pytest-instafail==0.5.0
pytest-pep8==1.0.6
pytest-timeout==2.1.0
pytest-twisted==1.14.0
pytest-xdist==3.2.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
robotframework==6.0.2
robotframework-requests==0.9.4
selenium==4.9.0
silx==1.1.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.23.0
virtualenv-clone==0.5.7
win-unicode-console==0.5
wsproto==1.2.0
zipp==3.15.0
